+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added conversion from `Vec<Feature>` to `GeoJson`.
+* Changed `Serialize` impls to avoid creating intermediate `JsonObject`s.
 
 ## 0.24.1